Use setState instead of re-rendering on each model update
I was misusing and misunderstanding React's rendering and lifecycle
model. Each app model update was constructing a new App instance, along
with its entire component subtree. This means that the PureRender
decorations aren't resulting in any performance improvement, since the
existing virtual DOM is being scrapped. Even worse, I think this was
actually re-rendering the DOM on every update, so even React's virtual
DOM diffing magic was being... bypassed. Woops!

The change is to have the model stream's updates call `setState` on the
root element. This lets React and the `PureRenderMixin` kick in with the
state and prop diffing. Since we're using an immutable `Record` for the
app state, and `props` for the descendent components, this should result
in only the dirty subtrees being re-rendered.