Add an experimental helper to reduce number of arrow functions
The helper is called `magicMap` at present, and helps with the common
case of mapping over the observable passed to the `dispatch` callback.

If this seems useful, I might attempt to make it into a full imitation
of the `Observable` API. The idea would be to build up a representation
of a pipeline of operations on an `Observable` that can be passed
directly as a callback for `dispatch`.